#275 [feat] 사용자 종료된 챌린지 갯수 로직 추가

+++ photi-core/domain/src/main/kotlin/com/photi/core/domain/user/dto/FindChallengeHistoryDto.kt
@@ -7,6 +7,6 @@ data class FindChallengeHistoryDto @QueryProjection constructor(
     val username: String,
     val imageUrl: String,
     val feedCount: Int,
-    val endedChallengeCount: Int,
+    val endedChallengeCount: Long,
     val signInDate: LocalDateTime,
 )
